Renames folders to match page names
